+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Reused the value component in the admin tag component of the admin control panel
+
 ### Changed
 
 - Upgraded `jsonpath` from version `1.1.1` to `1.2.1`